refactor: split out the advanced backend settings
The accordion body is its own component now: the three numeric fields, their
reset buttons and the log-from-other-modules checkbox. It takes the four
values as models, the file config and error lists as props, and emits which
field to reset, so it stays free of the validation library.

OnboardingSettings drops from 492 to 356 lines. The characterization spec
passes unchanged and is what proves the child is really rendered, since
components are not auto-imported here.
